+++ src/modules/goals/dtos/Sellers.dto.ts
@@ -1,0 +1,9 @@
+import { IsString } from 'class-validator';
+export class SellersDto {
+
+    @IsString()
+    co_ven: string;
+
+    @IsString()
+    des_ven: string;
+}